解释器

+++ chapter3-cf/classfile/attr_code.go
@@ -38,12 +38,12 @@ type ExceptionTableEntry struct {
 	catchType uint16
 }
 
-func (self *CodeAttribute) MaxLocals() uint16 {
-	return self.maxLocals
+func (self *CodeAttribute) MaxLocals() uint {
+	return uint(self.maxLocals)
 }
 
-func (self *CodeAttribute) MaxStack() uint16 {
-	return self.maxStack
+func (self *CodeAttribute) MaxStack() uint {
+	return uint(self.maxStack)
 }
 
 func (self *CodeAttribute) Code() []byte {
